Tests/LibWeb/Text/expected/caret-position-between-post-info-and-body.txt

caret below post info starts message: true
